Issues #3,#4,#5,#6 resolved

+++ team_headers/stopwatch.h
@@ -11,7 +11,7 @@
 typedef struct StopWatch
 {
 	Uint32 Start;
-	unsigned int Time;
+	unsigned long Time;
 }stopwatch_struct;
 
 void StopWatchSetUp(float time);
